chore(transitions): minor doc and code tweaks
