+++ src/modules/Lines/index.ts
@@ -11,6 +11,7 @@ import hoveredLineIndexFrag from '@/graph/modules/Lines/hovered-line-index.frag?
 import hoveredLineIndexVert from '@/graph/modules/Lines/hovered-line-index.vert?raw'
 import { defaultConfigValues } from '@/graph/variables'
 import { getCurveLineGeometry } from '@/graph/modules/Lines/geometry'
+import { updateAttributeBuffers } from '@/graph/modules/Shared/buffer'
 import { getBytesPerRow } from '@/graph/modules/Shared/texture-utils'
 import { ensureVec2, ensureVec4 } from '@/graph/modules/Shared/uniform-utils'
 import { readPixels } from '@/graph/helper'
@@ -594,7 +595,8 @@ export class Lines extends CoreModule {
     const { data } = this
     const linksNumber = data.linksNumber ?? 0
     const colorData = data.linkColors ?? new Float32Array(linksNumber * 4).fill(0)
-    const { source, target, previous } = this.updateAttributeBuffers(
+    const { source, target, previous } = updateAttributeBuffers(
+      this.device,
       colorData,
       this.sourceColorBuffer,
       this.targetColorBuffer,
@@ -617,7 +619,8 @@ export class Lines extends CoreModule {
     const { data } = this
     const linksNumber = data.linksNumber ?? 0
     const widthData = data.linkWidths ?? new Float32Array(linksNumber).fill(0)
-    const { source, target, previous } = this.updateAttributeBuffers(
+    const { source, target, previous } = updateAttributeBuffers(
+      this.device,
       widthData,
       this.sourceWidthBuffer,
       this.targetWidthBuffer,
@@ -1061,60 +1064,5 @@ export class Lines extends CoreModule {
       data: new Float32Array(4).fill(0),
     })
     this.linkStatusTextureSize = 0
-  }
-
-  private updateAttributeBuffers (
-    targetData: Float32Array,
-    sourceBuffer: Buffer | undefined,
-    targetBuffer: Buffer | undefined,
-    previousData: Float32Array | undefined,
-    tupleSize: 1 | 4
-  ): { source: Buffer; target: Buffer; previous: Float32Array } {
-    const oldCount = previousData ? previousData.length / tupleSize : 0
-    const newCount = targetData.length / tupleSize
-    const sameCount = oldCount === newCount
-
-    // Reuse both buffers when the topology is unchanged so the old target becomes the next source.
-    // TODO: Rare edge case - smooth in-flight attribute transitions when updates arrive mid-animation.
-    if (sameCount &&
-        sourceBuffer && !sourceBuffer.destroyed &&
-        targetBuffer && !targetBuffer.destroyed) {
-      const nextSource = targetBuffer
-      const nextTarget = sourceBuffer
-      nextTarget.write(targetData)
-      return {
-        source: nextSource,
-        target: nextTarget,
-        previous: new Float32Array(targetData),
-      }
-    }
-
-    const sourceData = new Float32Array(targetData.length)
-    const sharedCount = Math.min(oldCount, newCount)
-    for (let i = 0; i < sharedCount * tupleSize; i += 1) {
-      sourceData[i] = previousData?.[i] ?? targetData[i] ?? 0
-    }
-    for (let i = sharedCount * tupleSize; i < targetData.length; i += 1) {
-      sourceData[i] = targetData[i] ?? 0
-    }
-
-    if (sourceBuffer && !sourceBuffer.destroyed) {
-      sourceBuffer.destroy()
-    }
-    if (targetBuffer && !targetBuffer.destroyed) {
-      targetBuffer.destroy()
-    }
-
-    return {
-      source: this.device.createBuffer({
-        data: sourceData,
-        usage: Buffer.VERTEX | Buffer.COPY_DST,
-      }),
-      target: this.device.createBuffer({
-        data: targetData,
-        usage: Buffer.VERTEX | Buffer.COPY_DST,
-      }),
-      previous: new Float32Array(targetData),
-    }
   }
 }
